refactoring that partate model from receive
